Added Game Options
Cards can be face up or face down in each hand.
Added 5 card hands and Variable number of rounds.
Players can start with hands, and there is a method to allow players to choose the index of the card to donate to the wheel.
Requires some UI changes (How each hand is displayed, the ability to choose a card to donate to the wheel)